+++ packages/terra-framework-docs/CHANGELOG.md
@@ -7,6 +7,11 @@
 * Added
   * Added example and tests for `terra-worklist-data-grid` to select rows via mouse and keyboard.
 
+## 1.31.0 - (August 4, 2023)
+
+* Added
+  * Added examples and tests for `terra-worklist-data-grid` to interact with cells containing focusable elements.
+
 ## 1.30.0 - (August 2, 2023)
 
 * Changed